# Performance & Observability Epic: Response Time Monitoring & Optimization