… docs/FHIR_R4_Resources.csv
… first_request.json
… pyproject.toml
… src/nl_fhir/__pycache__/config.cpython-310.pyc
… src/nl_fhir/__pycache__/main.cpython-310.pyc
… src/nl_fhir/api/middleware/__pycache__/timing.cpython-310.pyc
… src/nl_fhir/services/__pycache__/conversion.cpython-310.pyc
… src/nl_fhir/services/__pycache__/monitoring.cpython-310.pyc
… src/nl_fhir/services/fhir/__pycache__/__init__.cpython-310.pyc
… src/nl_fhir/services/fhir/__pycache__/bundle_assembler.cpython-310.pyc
… src/nl_fhir/services/fhir/__pycache__/hapi_client.cpython-310.pyc
… src/nl_fhir/services/fhir/__pycache__/quality_optimizer.cpython-310.pyc
… src/nl_fhir/services/fhir/__pycache__/resource_factory.cpython-310.pyc
… src/nl_fhir/services/fhir/__pycache__/validator.cpython-310.pyc
… src/nl_fhir/services/fhir/bundle_assembler.py
… src/nl_fhir/services/fhir/hapi_client.py
… src/nl_fhir/services/fhir/quality_optimizer.py
… src/nl_fhir/services/fhir/resource_factory.py
… src/nl_fhir/services/fhir/validator.py
… src/nl_fhir/services/model_warmup.py
… src/nl_fhir/services/monitoring.py
… src/nl_fhir/services/nlp/__pycache__/entity_extractor.cpython-310.pyc
… src/nl_fhir/services/nlp/__pycache__/llm_processor.cpython-310.pyc
… src/nl_fhir/services/nlp/entity_extractor.py
… src/nl_fhir/services/nlp/llm_processor.py
… src/nl_fhir/services/task_workflow_service.py
… tests/test_conversion_vitals_extraction.py
… tests/test_model_warmup.py
… tests/test_observation_creation.py
… tests/test_performance_integration.py